chore: improve componentization of query result panel
