+++ src/app/(main)/query-runner/_components/cql-editor.tsx
@@ -18,14 +18,11 @@ import {
   ResizablePanelGroup,
 } from "@scylla-studio/components/ui/resizable";
 import { Skeleton } from "@scylla-studio/components/ui/skeleton";
-import { Tabs, TabsList, TabsTrigger } from "@scylla-studio/components/ui/tabs";
-import { useLayout } from "@scylla-studio/contexts/layout";
 import { useCqlFilters } from "@scylla-studio/hooks/use-cql-filters";
 import type { AvailableConnections } from "@scylla-studio/lib/connections";
 import type { TracingResult } from "@scylla-studio/lib/execute-query";
-import { cn } from "@scylla-studio/lib/utils";
 import debounce from "lodash.debounce";
-import { Braces, ChartArea, Play, SearchCode } from "lucide-react";
+import { Braces, Play } from "lucide-react";
 import type {
   CancellationToken,
   Position,
@@ -33,18 +30,10 @@ import type {
   languages,
 } from "monaco-editor";
 import { useAction } from "next-safe-action/hooks";
-import {
-  startTransition,
-  useCallback,
-  useEffect,
-  useRef,
-  useState,
-} from "react";
+import { useCallback, useEffect, useRef, useState } from "react";
 import { toast } from "sonner";
+import { CqlResultPanel } from "./cql-resultpanel";
 import { ResultFilters } from "./result-filters";
-import { ResultsRender } from "./results-render";
-import QueryDashboard from "./tracing-dashboard-render";
-import { TracingRender } from "./tracing-render";
 import {
   formatKeyspaces,
   formatKeyspacesTables,
@@ -54,12 +43,6 @@ import {
 enum ExecuteType {
   ALL = 0,
   CURRENT = 1,
-}
-
-enum DisplayTabs {
-  RESULT = "result",
-  TRACING = "tracing",
-  DASHBOARD = "dashboard",
 }
 
 export function CqlEditor() {
@@ -76,7 +59,6 @@ export function CqlEditor() {
     {} as TracingResult,
   );
   const [updateKey, setUpdateKey] = useState(0); // Used to force re-render when query is executed
-  const [activeTab, setActiveTab] = useState<string>(DisplayTabs.RESULT);
   const currentConnection = useCqlFilters((state) => state.currentConnection);
   const fetchSize = useCqlFilters((state) => state.fetchSize);
   // probably the editor functions uses a memo, so the instance of currentConnection is not updated
@@ -88,6 +70,7 @@ export function CqlEditor() {
   const editorReference = useRef<editor.IStandaloneCodeEditor | null>(null);
   const decorationsReference =
     useRef<editor.IEditorDecorationsCollection | null>(null);
+
   const queryExecutor = useAction(executeQueryAction, {
     onSuccess: ({ data }) => {
       setQueryResult(data?.result ?? []);
@@ -106,14 +89,6 @@ export function CqlEditor() {
       });
     },
   });
-
-  useEffect(() => {
-    currentConnectionReference.current = currentConnection ?? null;
-  }, [currentConnection]);
-
-  useEffect(() => {
-    fetchSizeReference.current = fetchSize ?? null;
-  }, [fetchSize]);
 
   const executeQuery = async (query: string) => {
     if (!currentConnectionReference.current) {
@@ -141,12 +116,6 @@ export function CqlEditor() {
     setLoadingResults(false);
   };
 
-  // Load saved query from localStorage when the component mounts
-  useEffect(() => {
-    const savedCode = localStorage.getItem("cqlEditorQuery");
-    if (savedCode) setCode(savedCode);
-  }, []);
-
   // Debounced save function using lodash
   // eslint-disable-next-line react-hooks/exhaustive-deps
   const debouncedSave = useCallback(
@@ -162,6 +131,7 @@ export function CqlEditor() {
     debouncedSave(updatedCode); // Save to local storage with debounce
   };
 
+  // Highlights line when put ";"
   const highlightQueryAtCursor = (
     editor: editor.IStandaloneCodeEditor,
     monaco: Monaco,
@@ -321,36 +291,21 @@ export function CqlEditor() {
     }
   };
 
-  const renderResult = useCallback(
-    () => (
-      <ResultsRender
-        key={`result-${updateKey}`}
-        data={queryResult}
-        tracingData={queryTracing}
-      />
-    ),
-    [queryTracing, updateKey],
-  );
+  useEffect(() => {
+    currentConnectionReference.current = currentConnection ?? null;
+  }, [currentConnection]);
 
-  const renderTracing = useCallback(
-    () => <TracingRender key={`tracing-${updateKey}`} data={queryTracing} />,
-    [queryTracing, updateKey],
-  );
-  const renderDashboard = useCallback(
-    () => (
-      <QueryDashboard
-        key={`dashboard-${updateKey}`}
-        tracingInfo={queryTracing}
-      />
-    ),
-    [queryTracing, updateKey],
-  );
-  const renderTabs = {
-    [DisplayTabs.RESULT]: renderResult(),
-    [DisplayTabs.TRACING]: renderTracing(),
-    [DisplayTabs.DASHBOARD]: renderDashboard(),
-  };
+  useEffect(() => {
+    fetchSizeReference.current = fetchSize ?? null;
+  }, [fetchSize]);
 
+  // Load saved query from localStorage when the component mounts
+  useEffect(() => {
+    const savedCode = localStorage.getItem("cqlEditorQuery");
+    if (savedCode) setCode(savedCode);
+  }, []);
+
+  // Load all keyspaces from the current connection
   useEffect(() => {
     if (currentConnection) queryKeySpaces(currentConnection);
   }, [currentConnection]);
@@ -409,56 +364,12 @@ export function CqlEditor() {
           )}
         </ResizablePanel>
         <ResizableHandle withHandle />
-        <ResizablePanel className="flex flex-row">
-          {!queryResult || queryResult.length === 0 ? null : (
-            <Tabs
-              orientation="vertical"
-              value={activeTab}
-              onValueChange={setActiveTab}
-              className="flex"
-            >
-              <TabsList className="flex h-full flex-col justify-start p-0 rounded-none">
-                <TabsTrigger
-                  value={DisplayTabs.RESULT}
-                  className={cn(
-                    "w-full justify-center rounded-none px-4 py-2 text-left hover:bg-white/10",
-                    activeTab === DisplayTabs.RESULT && "bg-white",
-                  )}
-                >
-                  <Braces size={16} />
-                </TabsTrigger>
-
-                <TabsTrigger
-                  value={DisplayTabs.TRACING}
-                  className={cn(
-                    "w-full justify-center rounded-none px-4 py-2 text-left hover:bg-white/10",
-                    activeTab === DisplayTabs.TRACING && "bg-white",
-                  )}
-                >
-                  <SearchCode size={18} />
-                </TabsTrigger>
-
-                <TabsTrigger
-                  value={DisplayTabs.DASHBOARD}
-                  className={cn(
-                    "w-full justify-center rounded-none px-4 py-2 text-left hover:bg-white/10",
-                    activeTab === DisplayTabs.DASHBOARD && "bg-white",
-                  )}
-                >
-                  <ChartArea size={18} />
-                </TabsTrigger>
-              </TabsList>
-            </Tabs>
-          )}
-          {loadingResults ? (
-            <div className="flex flex-col w-full h-full gap-1 p-1">
-              <Skeleton className="h-11 w-full" />
-              <Skeleton className="h-full w-full" />
-            </div>
-          ) : (
-            renderTabs[activeTab as DisplayTabs]
-          )}
-        </ResizablePanel>
+        <CqlResultPanel
+          updateKey={updateKey}
+          queryResult={queryResult}
+          queryTracing={queryTracing}
+          loadingResults={loadingResults}
+        />
       </ResizablePanelGroup>
     </div>
   );
